updatedeps: restructured files
