--- gulpfile.js
@@ -1,24 +1,0 @@
-var gulp = require("gulp");
-var ts = require("gulp-typescript");
-var uglify = require("gulp-uglify");
-var tsProject = ts.createProject("tsconfig.json");
-var pump = require("pump");
-var rename = require("gulp-rename");
-
-gulp.task("default", function() {
-  return tsProject
-    .src()
-    .pipe(tsProject())
-    .js.pipe(gulp.dest("dist"));
-});
-
-gulp.task("rename", function() {
-  return gulp
-    .src("dist/index.js")
-    .pipe(rename("index.min.js"))
-    .pipe(gulp.dest("dist"));
-});
-
-gulp.task("compress", gulp.series("rename"), function(cb) {
-  return pump([gulp.src("dist/*.min.js"), uglify(), gulp.dest("dist")], cb);
-});
