+++ tests/test.html
@@ -1,4 +1,4 @@
 <html>
     <script>var exports = {};</script>
-    <script src="../dist/index.js"></script>
+    <script src="./node_modules/typescript-string-operations/dist/index.min.js"></script>
 </html>
